[Justice Counts] Fix duplicate datapoints showing up on dashboards (Recidiviz/recidiviz-data#16591)
## Description of the change

Ivan noticed that on some of our charts, multiple values are showing up
for the same month. For instance, go to NCDPS on Staging and look at
Prisons:

![86551235-d256-4145-b462-d03e5f5e665b](https://user-images.githubusercontent.com/99899591/201183653-4078fc5a-9fb8-4dd2-b943-e246152112ae.png)

This is happening because in the database, we have a bunch of rows with
a deprecated disaggregated/dimension. Our `get_dimension` method fails
to find this disaggregation/dimension because it's no longer in our
codebase, and returns `None`. But that is also the value returned by
this method for an aggregate datapoint! So to the frontend's
perspecitive, aggregate datapoints look the same as deprecated
disaggregate datapoints.

To fix this, I'm adjusting `get_dimension` so it returns a tuple -- the
dimension and also a boolean indicating whether the dimension was
successfully parsed. This will help us distinguish from the case in
which the dimension is truly null vs. deprecated.

We will then only send datapoints to the frontend if we were able to
successfully parse their dimension.

To see how it looks fixed, I deployed this to playtesting and the same
chart looks much better:
<img width="1049" alt="Screen Shot 2022-11-10 at 2 06 06 PM"
src="https://user-images.githubusercontent.com/99899591/201184291-144cddae-57d9-47cb-9a14-badaf3c9d2ad.png">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#16590

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2af2a76dc399cb779ca875aba9b517b4ab44b6b8