[Justice Counts] Fix bug with deprecated datapoints (Recidiviz/recidiviz-data#16719)
## Description of the change

In https://github.com/Recidiviz/recidiviz-data/pull/16591 I introduced
two new pieces of logic:
* Filter out datapoints with deprecated dimensions after fetching from
the DB
* If we later see datapoints with deprecated dimensions somewhere in the
codebase, raise an error

However, this caused a bug because I missed a spot where I should be
filtering, so we were still seeing those deprecated datapoints later and
raising an error.

One way to fix is to add the filtering to the missed spot -- namely,
when we fetch reports with `include_datapoints=True`. However, when I
tried this, the filtering drastically slowed down report loading.

Instead, I'm actually going to remove one a piece of the error checking
that happens downstream, and instead just pass over datapoints that are
deprecated. This prevents us from checking for bad datapoints twice,
which was what was slowing things down badly.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9a900a2f7e5da7d972cdd255631a7f9d8025a2a5