[Justice Counts] put session.commit() into try catch when creating reports (Recidiviz/recidiviz-data#16701)
## Description of the change

When creating reports with the same start and end date, I was seeing a
generic response. The fix for this bug is two fold: This PR puts the
`session.commit()` call into the try/catch that throws the specific `A
report of that date range has already been created.` response because
the unique constraint exception is raised with `session.commit()` call.
The second half of this fix is in [this
PR](https://github.com/Recidiviz/justice-counts/compare/nichelle-hall/display-report-creation-error?expand=1),
that makes a change in the FE where we now check what the description is
before displaying the toast. We deprecated all api responses that were
not 500 because only 500 responses show up in our logs. As a result, the
check to see if the response code is 400 to display the toast with the
specific error date error was always ignored.

https://user-images.githubusercontent.com/19961693/202266767-2df7befa-6997-416a-9ad4-3ec29211642d.mov

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0f08aa8784acfeca43a31dae3a547a57be1dc22c