[Justice Counts] Create monthly and annual reports for agencies (if do not exist) (Recidiviz/recidiviz-data#16685)
## Description of the change

This PR sets the foundation for reports to be automatically generated
for each of our agencies at the start of every month and year. It does
so by:

1. Determining the current month and year.
2. Attempting to create a new monthly report for this agency and this
month. If a report already exists, do nothing.
3. If the month is January, attempting to create a new annual report for
this agency and this year. If a report already exists, do nothing.

Additional notes:
- When a report is created automatically, the user_account_id associated
with the report is set to None.
- A unit test was written to test 4 possible cases:
1. Monthly and Annual reports created
2. Only Monthly report created
3. Only Annual Report created
4. Neither Monthly nor Annual report created
- A unit test was written to test that a report was created when
user_account_id=None.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#15416

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 80e8620c2b7c0295acf62951d5f04fdadffd7a04