[Pathways] Remove supervision level override for ND supervision population snapshot (Recidiviz/recidiviz-data#16778)
## Description of the change

Removes the NULL supervision level override for ND in the
`pathways_deduped_supervision_sessions` view, which feeds
`supervision_population_snapshot_by_dimension` and
`supervision_population_time_series` views. We disable the supervission
level filter on the time series chart so these supervision levels are
effectively unreachable in the time series chart which is what we want.
But we do want to see them in the by dimension breakdown charts.

I have run this in the `jovergaag_11289_*` sandbox, exported it, and
loaded it into a FE for QA. It looks great!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#11289

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 74703d5f8fb5847daa39a120ceb7a8ef1fe1bc63