Create person_age_sessions and update age source in compartment_sessions (Recidiviz/recidiviz-data#16629)
## Description of the change

This PR is pre-work to a PR that creates
`compartment_sub_sessions_materialized` as part of Recidiviz/recidiviz-data#16628. Sub-sessions
will further sub-sessionize dataflow_sessions based on age. This PR does
not yet do that sub-sessionization, but does the first step in creating
`person_age_sessions` and using that as the source for age in
`compartment_sessions`.

There are a few minor discrepancies in compartment_sessions resulting in
the way that age was being calculated previously for which I think the
new approach is preferred. Previously we date-differenced a person's
birthdate with the session boundary and divided that by 365.25, which
can cause rounding issues. In a small % of cases, this causes the new
age calculated at a session boundary to be 1 year less than it was
originally (with the new version being more precise).

The following table (linked
[here](https://docs.google.com/spreadsheets/d/1iP5Bjb4VP1NJ4vASccpH1zGAAQ8p5F1rxQYLj_iN27Q/edit#gid=0)
with query) checks to make sure that any discrepancies in the age_start
and age_end values in compartment sessions are accounted for.

<img width="777" alt="Screen Shot 2022-11-21 at 2 59 48 PM"
src="https://user-images.githubusercontent.com/6909506/203174965-142a6bf4-f26f-421c-9af6-c7b1970382a6.png">

The new age values match the original values exactly ~99.9% of the time
and are within 1 year of the original 100% of the time.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 68d2263c0ff88c1bf42e2dcbadf516f7599f4383