[Justice Counts] Creating monthly and annual reports (if do not exist) based on custom freq (Recidiviz/recidiviz-data#16765)
## Description of the change

This PR completes Step 4 of
[Recidiviz/recidiviz-data#15416](https://app.zenhub.com/workspaces/justice-counts--622109ea3bb6e0001d2d06ad/issues/recidiviz/recidiviz-data/15416).
It builds on the work of previous PRs:
[Recidiviz/recidiviz-data#16675](https://github.com/Recidiviz/recidiviz-data/pull/16675) and
[Recidiviz/recidiviz-data#16685](https://github.com/Recidiviz/recidiviz-data/pull/16685).

Within report.create_new_reports(), we make two calls to
DatapointsForMetric.get_metric_definitions_for_report() - once for
monthly reports and once for annual reports.

DatapointsForMetric.get_metric_definitions_for_report() checks which
metrics should appear on the report of interest (monthly or annual) for
the current month. If there are no metric definitions for the report of
interest (monthly or annual) for the current month, we do not create a
report. If there are metric definitions for the report of interest
(monthly or annual) for the current month, we proceed to check if the
report of interest (monthly or annual) already exists. If the report of
interest (monthly or annual) does not already exist, one is created.

Note: Logging in create_new_reports.py has 3 cases for both monthly and
annual reports
1. The reports was created
2. The report was not created because it already exists
3. The report was not created because it is not enabled for the current
month

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#15416

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 41fa9c9253fa120918e0e46d8684f56cd5408051