[Workflows] Remove empty records from violations field (Recidiviz/recidiviz-data#16855)
## Description of the change

When there were no violations to put in the metadata notes, the SQL
query was creating an array with a single empty object (`[{}]`) instead
of an empty array. When the front-end tried to parse this information to
display, it threw an error on the empty object and didn't display the
opportunity for the client. This PR adds a check to remove that empty
array if it exists.

Tested by exporting this query into staging and seeing that someone who
should have been eligible for SLD now renders as such.

This was affecting the majority of SLD-eligible clients.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#16838

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- ~[ ] Tests have been written to cover the code changed/added as part
of this pull request~ BQ change

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1e9a435fc6454d166cfc4532eceb928a0fd5c537